requirements.txt
Flask
mysql-connector-python
pytz
requests
firebase_admin
reportlab
qrcode[pil]
pillow